rename 'instance_to_array'-like files to 'to_array_debug'-like
