+++ runtime/job-workers/job-interface.h
@@ -8,7 +8,7 @@
 #include "common/wrappers/string_view.h"
 
 #include "runtime/instance-copy-processor.h"
-#include "runtime/instance-to-array-processor.h"
+#include "runtime/to-array-processor.h"
 #include "runtime/kphp_core.h"
 #include "runtime/refcountable_php_classes.h"
