+++ tests/phpt/cl/005_cast_to_array_debug.php
