+++ tests/phpt/cl/007_to_array_debug_with_extends.php
