+++ tests/phpt/cl/009_to_array_debug_with_class_names.php
