+++ tests/phpt/class_inheritance/010_to_array_debug.php
