+++ tests/phpt/ffi/004_to_array_debug_error.php
